+++ tests/testthat/test-full_seq.R
@@ -1,5 +1,3 @@
-context("full_seq")
-
 test_that("full_seq errors if sequence isn't regular", {
   expect_error(full_seq(c(1, 3, 4), 2), "not a regular sequence")
   expect_error(full_seq(c(0, 10, 20), 11, tol = 1.8), "not a regular sequence")